Socket messages without header
To interoperate with Rationale Python code, it is easier to send socket messages without the built-in header. Instead, messages have a suffix that lets the receiver know the cutoff point of each message. Any extra bytes are kept in memory until the next message. This is because the Python code receives a fixed number of bytes, but messages may be any length.